Updating for SS4
- move code to src
- add namespace, split out Readonly fields to own files
- run updater script
- update composer.json
